+++ resale/src/main/java/com/goti/resale/service/domain/ResaleOrderDomainService.java
@@ -1,0 +1,22 @@
+package com.goti.resale.service.domain;
+
+import java.util.List;
+import java.util.UUID;
+
+import com.goti.domain.vo.TransactionItemVO;
+import com.goti.resale.domain.entity.resale.ResaleHoldEntity;
+import com.goti.resale.domain.entity.resale.ResaleRestrictionEntity;
+import com.goti.resale.dto.response.ResaleOrderCreateResponse;
+
+public interface ResaleOrderDomainService {
+	String generateOrderNumber();
+
+	String generateResaleTicketNumber(String resaleSuffix, String num);
+
+	List<TransactionItemVO> calculateOrderItems(UUID buyerId, List<ResaleHoldEntity> holds,
+		ResaleRestrictionEntity restriction);
+
+	void validatePossessionLimit(int currentOwnedCount, int pendingCount, int requestCount);
+
+	ResaleOrderCreateResponse initOrder(UUID buyerId, List<ResaleHoldEntity> holds, UUID gameId);
+}
